+++ playgrounds/demo/app/composables/usePageMeta.ts
@@ -61,6 +61,7 @@ const components: IComponentInfo[] = [
   { id: 'description-list', description: 'For cases where you need to convert data from one row of a table into a separate table to make the page look more complete.', iconData: GroupIcon },
   { id: 'dropdown-menu', description: 'A menu to display actions when clicking on an element.', iconData: GroupIcon },
   { id: 'context-menu', description: 'A context menu component.', iconData: GroupIcon },
+  { id: 'empty', description: 'An empty state component.', iconData: GroupIcon },
   { id: 'field-group', description: 'Group multiple button-like elements together.', iconData: GroupIcon },
   { id: 'file-upload', description: 'A file upload input component.', iconData: FormIcon },
   { id: 'form', description: 'A form component with built-in validation and submission handling.', iconData: FormSettingsIcon },